Refactore codebase and introduce Environment class
This commit marks significant codebase improvement facilitating both code maintainability and readability. Seeing major changes across multiple files, the introduction of Environment class remained remarkable.
Following are the elaborated changes:
- Added white spaces in K8sConstants.kt improving code readability and clarity.
- The REGISTRY_DEFAULT_PORT constant was discarded due to its non-usage.
- Enhanced code formatting and organization in Container.kt and ContainerTaskTests.kt.
- Introduced an additional empty line in ContainerLogStreamer.kt and eradicated irrelevant imports from K8sJobRuntime.kt.
- Revised the indentation and rectified the method utilized to remove the container in files ContainerTaskTests.kt , ImageBuilderTests.kt, and ContainerLifeCycleTests.kt.
- Substantive alterations visible in DockerConstants.kt and DockerClientFactory.kt. Various constants and methods have been adjusted relevant to Docker daemon endpoint and mode of network.
- Redressed ImageBuilderTests.kt to accommodate shifts in Docker-related constants.

An Environment class was introduced in this commit to move the management of environment variables and system properties that the application relies on into a central place, giving a single source of truth. By doing so, it allows for better maintainability and easier changes or extension in the future. Encouraging the use of constants over string literals for environment variables decreases the chances of mistakes, hence making the code robust.